pyproject.toml
[tool.poetry]
name = "homie-backend"
version = "0.1.0"
description = "Homie app backend"
authors = ["Daniel Kiermeier <kiermeier.dan@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
debugpy = "^1.6.7"
django-debug-toolbar = "^4.2.0"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.27"
django-q2 = "^1.6.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
combine_as_imports = true